fix logging database in file
