+++ backend/src/Makefile
@@ -8,7 +8,7 @@ all: rebuild check
 rebuild: socket_server json_export_test clean
 
 socket_server: clean socket_server.c
-	$(CC) $(CFLAGS) $(EXTRAFLAGS) socket_server.c socket_util.c json_export.c ../../data/src/db_api.c ../../data/src/db_logger.c common/logger.c -o $(BUILD_PATH)socket_server  -lpthread -lmysqlclient
+	$(CC) $(CFLAGS) $(EXTRAFLAGS) socket_server.c socket_util.c json_export.c ../../data/src/db_api.c common/logger.c -o $(BUILD_PATH)socket_server  -lpthread -lmysqlclient
 
 json_export_test: clean tests/json_export_test.c
 	$(CC) $(CFLAGS) $(EXTRAFLAGS) tests/json_export_test.c json_export.c -o $(BUILD_PATH)json_export_test
