+++ backend/src/common/logger.c
@@ -1,14 +1,17 @@
 #include "logger.h"
-#include <stdlib.h>
-#include <time.h>
 
 static FILE *plog_file = NULL;
+static FILE *fot_db_log = NULL;
 
 void init_logger(void) {
     plog_file = fopen("log.txt", "a");
+    fot_db_log = fopen("db_log.txt", "a");
     if (plog_file == NULL) {
             fprintf(stderr, "Can't open log file\n");
             exit(EXIT_FAILURE);
+    }
+    if (fot_db_log == NULL) {
+        fprintf(stderr, "Can't open log file\n");
     }
     fprintf(plog_file, "-------------------\n");
 }
@@ -50,4 +53,27 @@ void print_date_time(void) {
         fprintf(plog_file, "%d-%02d-%02d %02d:%02d:%02d | ", timeinfo->tm_year + 1900, timeinfo->tm_mon + 1,
                 timeinfo->tm_mday, timeinfo->tm_hour, timeinfo->tm_min, timeinfo->tm_sec);
     }
+}
+
+void db_close_logger(void) {
+    if (fot_db_log != NULL) {
+        fclose(fot_db_log);
+    }
+    fot_db_log = NULL;
+}
+
+void db_log_message(const char *type, const char *message) {
+    if (fot_db_log == NULL) {
+        init_logger();
+    }
+    if (fot_db_log != NULL) {
+        time_t current_time = time(NULL);
+        struct tm *now = localtime(&current_time);
+        fprintf(fot_db_log, "[DB] ");
+        fprintf(fot_db_log, "[%02d.%02d.%d ", now->tm_mday, now->tm_mon + 1, now->tm_year + 1900);
+        fprintf(fot_db_log, "%02d:%02d:%02d] ", now->tm_hour, now->tm_min, now->tm_sec);
+        fprintf(fot_db_log, "[%s] %s\n", type, message);
+        fflush(fot_db_log);
+    }
+    db_close_logger();
 }
