+++ docker-compose.yml
@@ -14,7 +14,7 @@ services:
     build:
       context: ./
       args:
-        - MYSQL_DATABASE=${MYSQL_DATABASE} 
+        - MYSQL_DATABASE=${MYSQL_DATABASE}
         - MYSQL_USER=${MYSQL_USER}
         - MYSQL_PASSWORD=${MYSQL_PASSWORD}
         - MYSQL_ROOT_PASSWORD=${MYSQL_ROOT_PASSWORD}
@@ -36,32 +36,32 @@ services:
     build:    
       context: ./data/
       args:
-        - MYSQL_DATABASE=${MYSQL_DATABASE} 
+        - MYSQL_DATABASE=${MYSQL_DATABASE}
         - MYSQL_USER=${MYSQL_USER}
         - MYSQL_PASSWORD=${MYSQL_PASSWORD}
         - MYSQL_ROOT_PASSWORD=${MYSQL_ROOT_PASSWORD}
         - MYSQL_ROOT_HOST=${MYSQL_ROOT_HOST}
         - MYSQL_PORT=${MYSQL_PORT}
         - MYSQL_INITDB_SKIP_TZINFO=${MYSQL_INITDB_SKIP_TZINFO}
-    env_file: 
-      - .env 
+    env_file:
+      - .env
     # NOTE: use of "mysql_native_password" is not recommended: https://dev.mysql.com/doc/refman/8.0/en/upgrading-from-previous-series.html#upgrade-caching-sha2-password
     # (this is just an example, not intended to be a production configuration)
     # command: mysqld --default-authentication-plugin=caching_sha2_password
     # указывает на то что контейнер будет перезапущен при сбое или перезагрузке сервера
     restart: always
     expose:
-      - "3306" 
+      - "3306"
       - "33060"
     networks:
       - db-network
 
     # прописываем общую директорию для того что бы при перезапуске контейнера сохранились данные базы данных
-    volumes:    
+    volumes:
         - ./data/docker/db_data/:/var/lib/mysql
 
 networks:
-  db-network: 
+  db-network:
     driver: bridge
     external: false
   frontend-network: {}
